+++ CHANGELOG.md
@@ -22,6 +22,7 @@ FEATURES
 - **New Resource:** Service Notification
 - **New Resource:** Service Binding
 - **New Resource:** Service Users
+- **New Resource:** Service Role
 - **New Resource:** Credential
 
 ### Serverless